+++ src/browser.js
@@ -1,3 +1,5 @@
+const axios = require('axios/dist/browser/axios.cjs')
+
 /**
  * Platform dependant utilities for the openEO JS Client.
  * 
@@ -15,6 +17,16 @@ class Environment {
 	 */
 	static getName() {
 		return 'Browser';
+	}
+
+	/**
+	 * Returns the axios client.
+	 * 
+	 * @returns {axios}
+	 * @static
+	 */
+	static axios() {
+		return axios;
 	}
 
 	/**