+++ src/connection.js
@@ -1,7 +1,6 @@
 const Environment = require('./env');
 const Utils = require('@openeo/js-commons/src/utils');
 const ProcessRegistry = require('@openeo/js-commons/src/processRegistry');
-const axios = require('axios').default;
 const StacMigrate = require('@radiantearth/stac-migrate');
 
 const AuthProvider = require('./authprovider');
@@ -1266,6 +1265,7 @@ class Connection {
 			options.signal = abortController.signal;
 		}
 
+		const axios = Environment.axios();
 		try {
 			let response = await axios(options);
 			let capabilities = this.capabilities();