+++ src/openeo.js
@@ -1,4 +1,3 @@
-const axios = require('axios').default;
 const Utils = require('@openeo/js-commons/src/utils');
 const Versions = require('@openeo/js-commons/src/versions');
 
@@ -52,6 +51,7 @@ class OpenEO {
 		let wellKnownUrl = Utils.normalizeUrl(url, '/.well-known/openeo');
 		let versionedUrl = url;
 		let response = null;
+		const axios = OpenEO.Environment.axios();
 		try {
 			response = await axios.get(wellKnownUrl, {timeout: 5000});
 
@@ -109,7 +109,7 @@ class OpenEO {
 	 * @returns {string} Version number (according to SemVer).
 	 */
 	static clientVersion() {
-		return "2.5.1";
+		return "2.5.2";
 	}
 
 }