+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.5.2] - 2024-07-11
 
 ### Added
 
@@ -15,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `usage`, `log_level`, `links` in batch jobs and services
 - Support for `level` parameter for log requests in batch jobs and services
 - Support passing additional parameters to synchronous jobs.
+
+### Changed
+
+- Updated axios from 0.x to 1.x
 
 ### Removed
 
@@ -211,7 +218,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 All prior releases have been documented in the [GitHub Releases](https://github.com/Open-EO/openeo-js-client/releases).
 
-[Unreleased]: https://github.com/Open-EO/openeo-js-client/compare/v2.5.1...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-js-client/compare/v2.5.2...HEAD
+[2.5.2]: https://github.com/Open-EO/openeo-js-client/compare/v2.5.1...v2.5.2
 [2.5.1]: https://github.com/Open-EO/openeo-js-client/compare/v2.5.0...v2.5.1
 [2.5.0]: https://github.com/Open-EO/openeo-js-client/compare/v2.4.1...v2.5.0
 [2.4.1]: https://github.com/Open-EO/openeo-js-client/compare/v2.4.0...v2.4.1